+++ package.json
@@ -8,8 +8,10 @@
     "@testing-library/user-event": "^7.2.1",
     "axios": "^0.20.0",
     "formik": "^2.2.0",
+    "qs": "^6.9.4",
     "react": "^16.14.0",
     "react-dom": "^16.14.0",
+    "react-hook-form": "^6.9.4",
     "react-redux": "^7.2.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
