--- src/components/Information/Information.module.css
@@ -1,85 +1,0 @@
-.container {
-  /* border: 1px solid black; */
-  padding: 5rem 20rem;
-}
-
-.title {
-  /* border: 1px solid black; */
-  font-size: 3rem;
-  font-weight: 700;
-  border-bottom: 1px solid #9f9f9f;
-  padding: 2.5rem 0;
-  margin: 0 0 2rem 0;
-}
-
-.form {
-  /* border: 1px solid black; */
-  display: flex;
-  justify-content: left;
-  flex-direction: column;
-  width: 50%;
-}
-
-.mandatory {
-  color: red;
-}
-
-.desc {
-  font-size: 1.4rem;
-  font-weight: bold;
-  margin: 2rem 0;
-}
-
-.input1 {
-  border: none;
-  background-color: #fcfcfc;
-  border-bottom: 1px solid #9f9f9f;
-  height: 4rem;
-  font-size: 1.4rem;
-}
-.input2 {
-  border: none;
-  background-color: #fcfcfc;
-  border-bottom: 1px solid #1d94a8;
-  height: 4rem;
-  font-size: 1.4rem;
-}
-.input3 {
-  border: none;
-  background-color: #fcfcfc;
-  border-bottom: 1px solid #1d94a8;
-  height: 23rem;
-  font-size: 1.4rem;
-  position: relative;
-  left: 0;
-  top: 0;
-}
-
-.alert {
-  color: red;
-  font-weight: 500;
-}
-
-.anonym {
-  font-weight: 400;
-  font-size: 1.4rem;
-}
-
-.checkbox {
-  background-color: #f4f4f4;
-  border-radius: 4px;
-  border: 1px solid #9f9f9f;
-  width: 1.6rem;
-  height: 1.6rem;
-}
-
-.optional {
-  font-weight: 400;
-  font-style: italic;
-  font-size: 1.4rem;
-}
-
-.check {
-  margin: 1rem 0;
-  display: flex;
-}
